- feat : feign
 - feign : @builder 어노테이션 추가

 - feign config : FeignConfig 재시도 로직 추가

+++ review-service/src/main/java/com/timebank/review/application/dto/ReviewDto.java
@@ -17,7 +17,7 @@ public class ReviewDto {
 	private Long transactionId;
 	private Long reviewerId;
 	private Long revieweeId;
-	private Byte rating;
+	private int rating;
 	private String comment;
 
 	public static ReviewDto fromEntity(Review review) {
